refactor : 딜러와 플레이어의 중복되는 부분을 추상 메서드로 분리
