+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -2,17 +2,14 @@ package blackjack.domain.participant;
 
 import blackjack.domain.carddeck.Card;
 import blackjack.domain.carddeck.CardDeck;
-import java.util.Collections;
-import java.util.List;
 
-public class Dealer {
+public class Dealer extends Participant {
 
     private static final int LIMIT_SCORE = 17;
-    private final Hand hand;
+
     private final CardDeck cardDeck;
 
     public Dealer() {
-        this.hand = new Hand();
         this.cardDeck = CardDeck.newShuffledDeck();
     }
 
@@ -20,19 +17,8 @@ public class Dealer {
         return cardDeck.draw();
     }
 
-    public void receiveCard(final Card card) {
-        this.hand.addCard(card);
-    }
-
+    @Override
     public boolean isOverLimitScore() {
         return getTotalScore() >= LIMIT_SCORE;
-    }
-
-    public int getTotalScore() {
-        return hand.totalScore();
-    }
-
-    public List<Card> getHand() {
-        return Collections.unmodifiableList(hand.toList());
     }
 }
