+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,0 +1,28 @@
+package blackjack.domain.participant;
+
+import blackjack.domain.carddeck.Card;
+import java.util.Collections;
+import java.util.List;
+
+public abstract class Participant {
+
+    private final Hand hand;
+
+    protected Participant() {
+        this.hand = new Hand();
+    }
+
+    public abstract boolean isOverLimitScore();
+
+    public void receiveCard(final Card card) {
+        this.hand.addCard(card);
+    }
+
+    public int getTotalScore() {
+        return hand.totalScore();
+    }
+
+    public List<Card> getHand() {
+        return Collections.unmodifiableList(hand.toList());
+    }
+}
